[201911] DellEMC S6100: Update SSD upgrade status checker (#8225)
Why I did it
To handle newer SSD firmware version in DellEMC S6100 platform (S210506G - 3IE devices).

How I did it
Update s6100_ssd_upgrade_status.sh to handle newer SSD firmware version.

How to verify it
Logs: UT_logs.txt

+++ platform/broadcom/sonic-platform-modules-dell/s6100/scripts/s6100_ssd_upgrade_status.sh
@@ -50,13 +50,13 @@ if [ $SSD_UPGRADE_STATUS1 == "2" ]; then
 
     echo "$0 `date` Upgraded to unknown version after first mp_64 upgrade." >> $SSD_UPGRADE_LOG
 
-elif [ $SSD_UPGRADE_STATUS2 == "2" ];then
+elif [ $SSD_MODEL == "3IE3" ] && [ $SSD_UPGRADE_STATUS2 == "2" ];then
     rm -rf $SSD_FW_UPGRADE/GPIO7_*
     touch $SSD_FW_UPGRADE/GPIO7_error
 
     echo "$0 `date` Upgraded to unknown version after second mp_64 upgrade." >> $SSD_UPGRADE_LOG
 
-elif [ $SSD_FW_VERSION == "s141002g" ] || [ $SSD_FW_VERSION == "s16425cg" ]; then
+elif [ $SSD_FW_VERSION == "s210506g" ] || [ $SSD_FW_VERSION == "s16425cg" ]; then
     # If SSD Firmware is upgraded
     GPIO_STATUS=$(echo "$iSMART_CMD" | grep GPIO | awk '{print $NF}')
 
@@ -75,11 +75,11 @@ elif [ $SSD_FW_VERSION == "s141002g" ] || [ $SSD_FW_VERSION == "s16425cg" ]; the
             systemctl start --no-block s6100-ssd-monitor.timer
 
             if [ $SSD_MODEL  == "3IE" ];then
-                echo "$0 `date` SSD FW upgraded from S141002C to S141002G in first mp_64." >> $SSD_UPGRADE_LOG
+                echo "$0 `date` SSD FW upgraded from S141002C to S210506G." >> $SSD_UPGRADE_LOG
             else
                 echo "$0 `date` SSD FW upgraded from S16425c1 to S16425cG in first mp_64." >> $SSD_UPGRADE_LOG
             fi
-        elif [ $SSD_UPGRADE_STATUS2 == "1" ]; then
+        elif [ $SSD_MODEL == "3IE3" ] && [ $SSD_UPGRADE_STATUS2 == "1" ]; then
             rm -rf $SSD_FW_UPGRADE/GPIO7_*
             touch $SSD_FW_UPGRADE/GPIO7_low
             logger -p user.crit -t DELL_S6100_SSD_MON "The SSD on this unit is faulty. Do not power-cycle/reboot this unit!"
@@ -88,14 +88,6 @@ elif [ $SSD_FW_VERSION == "s141002g" ] || [ $SSD_FW_VERSION == "s16425cg" ]; the
             echo "$0 `date` SSD entered loader mode in first mp_64 and upgraded to latest version after second mp_64." >> $SSD_UPGRADE_LOG
         fi
     fi
-
-elif [ $SSD_FW_VERSION == "s210506g" ] || [ $SSD_UPGRADE_STATUS2 == "3" ]; then
-    rm -rf $SSD_FW_UPGRADE/GPIO7_*
-    touch $SSD_FW_UPGRADE/GPIO7_low
-    logger -p user.crit -t DELL_S6100_SSD_MON "The SSD on this unit is faulty. Do not power-cycle/reboot this unit!"
-    logger -p user.crit -t DELL_S6100_SSD_MON "soft-/fast-/warm-reboot is allowed."
-
-    echo "$0 `date` SSD entered loader mode in first mp_64 and upgraded to version s210506g after second mp_64." >> $SSD_UPGRADE_LOG
 
 else
     if [ $SSD_UPGRADE_STATUS1 == "ff" ] && [ $SSD_UPGRADE_STATUS2 == "ff" ]; then
@@ -112,7 +104,7 @@ else
 
         echo "$0 `date` SSD entered loader mode in first mp_64 upgrade." >> $SSD_UPGRADE_LOG
 
-        if [ $SSD_UPGRADE_STATUS2 == "0" ]; then
+        if [ $SSD_MODEL == "3IE3" ] && [ $SSD_UPGRADE_STATUS2 == "0" ]; then
             echo "$0 `date` SSD entered loader mode in first mp_64 and recovered back to older version in second mp_64." >> $SSD_UPGRADE_LOG
         fi
     fi
